Feature/liquibase setup (#5)
* refactor: adjust types to better fit use case

* feat: implement liquibase, add existing entities, create mock data

* chore: add real symbols and images, fix foreign key constraints

* chore: yaml -> xml changelog

---------

+++ src/main/kotlin/pl/umcs/picto3/round/Round.kt
@@ -33,10 +33,10 @@ data class Round(
     val selectedSymbols: List<Symbol> = emptyList(),
 
     @Column(name = "speaker_response_time_ms")
-    val speakerResponseTime: Short?,
+    val speakerResponseTime: Int?,
 
     @Column(name = "listener_response_time_ms")
-    val listenerResponseTime: Short?,
+    val listenerResponseTime: Int?,
 
     @CreationTimestamp
     @Column(name = "started_at")
